+++ BilibiliLiveRecordDownLoader/Services/DownloadTaskPool.cs
@@ -20,7 +20,7 @@ namespace BilibiliLiveRecordDownLoader.Services
         /// <summary>
         /// record 开始/停止下载
         /// </summary>
-        public async Task DownloadAsync(LiveRecordListViewModel record, string path, double threads)
+        public async Task DownloadAsync(LiveRecordListViewModel record, string path, ushort threads)
         {
             var id = record.Rid;
             var startTime = record.StartTime;
